Initial Commit
First working try. It senses objects.
